+++ build.js
@@ -3,6 +3,11 @@
 const jetpack = require('fs-jetpack'),
     Terser  = require('terser');
 
+
+
+const buildDate = new Date().toISOString().split('T')[0],
+    buildID = require('git-commit-id')();
+
 function build () {
     const jsFiles = jetpack.find('./scripts', {
         matching : '*.js',
@@ -17,7 +22,7 @@ function build () {
 
         let version = source.match(/(v|version\s+)(\d+\.\d+.\d+)/);
         if (version == null) {
-            version = '0.0.1';
+            version = '0.0.0';
             console.log(`missing version info: ${file}`);
         } else {
             version = version[2];
@@ -32,7 +37,7 @@ function build () {
         path.push(name);
         path = path.join('/');
         
-        const ret = `// ${name}@v${version}, for SugarCube 2, by Chapel\n;${result.code}\n// end ${name}`;
+        const ret = `// ${name}, for SugarCube 2, by Chapel\n// v${version}, ${buildDate}, ${buildID}\n;${result.code}\n// end ${name}`;
         
         jetpack.write(path, ret, {atomic : true});
     });